+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 Changelog
 =========
+## Unreleased
+### Added
+- `UrlRule` to assist with prefixing rules (such as a version)
+
 ## 2.0.0-rc.1 - 2018-01-15
 ### Changed
 - Class namespaces include 'craft' to avoid future conflicts